+++ native_library/apple/FacebookANE/FacebookANE/SwiftController.swift
@@ -26,9 +26,9 @@ public class SwiftController: NSObject, SharingDelegate {
     public var context: FreContextSwift!
     public var functionsToSet: FREFunctionMap = [:]
     internal var appDidFinishLaunchingNotification: Notification?
-    internal var onShareSuccessEventId: String?
-    internal var onShareCancelEventId: String?
-    internal var onShareErrorEventId: String?
+    internal var onShareSuccessCallbackId: String?
+    internal var onShareCancelCallbackId: String?
+    internal var onShareErrorCallbackId: String?
     private var shareDialogs = [String: ShareDialog]()
     private var messageDialogs = [String: MessageDialog]()
     private var shareAPIs = [String: ShareAPI]()
@@ -173,9 +173,9 @@ public class SwiftController: NSObject, SharingDelegate {
     func login(ctx: FREContext, argc: FREArgc, argv: FREArgv) -> FREObject? {
         guard argc > 4,
             let permissions = [String](argv[0]),
-            let onSuccessEventId = String(argv[2]),
-            let onCancelEventId = String(argv[3]),
-            let onErrorEventId = String(argv[4])
+            let onSuccessCallbackId = String(argv[2]),
+            let onCancelCallbackId = String(argv[3]),
+            let onErrorCallbackId = String(argv[4])
             else {
                 return FreArgError(message: "login").getError(#file, #line, #column)
         }
@@ -184,11 +184,11 @@ public class SwiftController: NSObject, SharingDelegate {
                            viewController: UIApplication.shared.keyWindow?.rootViewController) { loginResult in
             switch loginResult {
             case .failed(let error):
-                self.dispatchLoginEvent(name: FacebookEvent.ON_LOGIN_ERROR, eventId: onErrorEventId, error: error, data: nil)
+                self.dispatchLoginEvent(name: FacebookEvent.ON_LOGIN_ERROR, callbackId: onErrorCallbackId, error: error, data: nil)
             case .cancelled:
-                self.dispatchLoginEvent(name: FacebookEvent.ON_LOGIN_CANCEL, eventId: onCancelEventId, error: nil, data: nil)
+                self.dispatchLoginEvent(name: FacebookEvent.ON_LOGIN_CANCEL, callbackId: onCancelCallbackId, error: nil, data: nil)
             case .success(let grantedPermissions, let declinedPermissions, let accessToken):
-                self.dispatchLoginEvent(name: FacebookEvent.ON_LOGIN_SUCCESS, eventId: onSuccessEventId, error: nil,
+                self.dispatchLoginEvent(name: FacebookEvent.ON_LOGIN_SUCCESS, callbackId: onSuccessCallbackId, error: nil,
                                         data: ["accessToken": accessToken.toDictionary(),
                                                "recentlyGrantedPermissions": grantedPermissions.compactMap { $0.name },
                                                "recentlyDeniedPermissions": declinedPermissions.compactMap { $0.name }])
@@ -197,9 +197,9 @@ public class SwiftController: NSObject, SharingDelegate {
         return nil
     }
     
-    func dispatchLoginEvent(name: String, eventId: String, error: Error?, data: [String: Any]?) {
+    func dispatchLoginEvent(name: String, callbackId: String, error: Error?, data: [String: Any]?) {
         var props = [String: Any]()
-        props["eventId"] = eventId
+        props["callbackId"] = callbackId
         if let err = error {
             props["data"] = ["message": err.localizedDescription]
         }
@@ -270,19 +270,19 @@ public class SwiftController: NSObject, SharingDelegate {
     
     func refreshCurrentAccessTokenAsync(ctx: FREContext, argc: FREArgc, argv: FREArgv) -> FREObject? {
         guard argc > 1,
-            let onTokenRefreshedEventId = String(argv[0]),
-            let onTokenRefreshFailedEventId = String(argv[1])
+            let onTokenRefreshedCallbackId = String(argv[0]),
+            let onTokenRefreshFailedCallbackId = String(argv[1])
             else {
                 return FreArgError(message: "setDefaultAudience").getError(#file, #line, #column)
         }
         AccessToken.refreshCurrentAccessToken { _, _, error in
             var props = [String: Any]()
             if let err = error {
-                props["eventId"] = onTokenRefreshFailedEventId
+                props["callbackId"] = onTokenRefreshFailedCallbackId
                 props["data"] = ["message": err.localizedDescription]
                 self.dispatchEvent(name: FacebookEvent.ON_TOKEN_REFRESH_FAILED, value: JSON(props).description)
             } else {
-                props["eventId"] = onTokenRefreshedEventId
+                props["callbackId"] = onTokenRefreshedCallbackId
                 if let accessToken = AccessToken.current {
                     props["data"] = accessToken.toDictionary()
                 }
@@ -423,17 +423,17 @@ public class SwiftController: NSObject, SharingDelegate {
     }
     
     func onShareSuccess(ctx: FREContext, argc: FREArgc, argv: FREArgv) -> FREObject? {
-        onShareSuccessEventId = String(argv[0])
+        onShareSuccessCallbackId = String(argv[0])
         return nil
     }
     
     func onShareCancel(ctx: FREContext, argc: FREArgc, argv: FREArgv) -> FREObject? {
-        onShareCancelEventId = String(argv[0])
+        onShareCancelCallbackId = String(argv[0])
         return nil
     }
     
     func onShareError(ctx: FREContext, argc: FREArgc, argv: FREArgv) -> FREObject? {
-        onShareErrorEventId = String(argv[0])
+        onShareErrorCallbackId = String(argv[0])
         return nil
     }